+++ LLDebugTool/DebugTool/LLDebugTool.h
@@ -65,7 +65,15 @@
  */
 - (void)stopWorking;
 
+/**
+ Whether working or not.
+ */
 @property (nonatomic , assign , readonly) BOOL isWorking;
+
+/**
+ LLDebugTool's version.
+ */
+@property (nonatomic , copy , readonly) NSString *version;
 
 /**
  Automatic open debug view controller with index.