+++ LLDebugTool/UserInterface/Tool/LLTool.m
@@ -72,6 +72,10 @@ static LLTool *_instance = nil;
     return [self.staticDateFormatter dateFromString:string];
 }
 
+- (NSString *)staticStringFromDate:(NSDate *)date {
+    return [self.staticDateFormatter stringFromDate:date];
+}
+
 + (UIView *)lineView:(CGRect)frame superView:(UIView *)superView {
     UIView *view = [[UIView alloc] initWithFrame:frame];
     view.backgroundColor = [UIColor lightGrayColor];
@@ -82,6 +86,25 @@ static LLTool *_instance = nil;
         [superView addSubview:view];
     }
     return view;
+}
+
++ (NSString *)prettyJSONStringFromData:(NSData *)data
+{
+    if ([data length] == 0) {
+        return nil;
+    }
+    NSString *prettyString = nil;
+    
+    id jsonObject = [NSJSONSerialization JSONObjectWithData:data options:0 error:NULL];
+    if ([NSJSONSerialization isValidJSONObject:jsonObject]) {
+        prettyString = [[NSString alloc] initWithData:[NSJSONSerialization dataWithJSONObject:jsonObject options:NSJSONWritingPrettyPrinted error:NULL] encoding:NSUTF8StringEncoding];
+        // NSJSONSerialization escapes forward slashes. We want pretty json, so run through and unescape the slashes.
+        prettyString = [prettyString stringByReplacingOccurrencesOfString:@"\\/" withString:@"/"];
+    } else {
+        prettyString = [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
+    }
+    
+    return prettyString;
 }
 
 #pragma mark - Lazy load