+++ LLDebugToolDemo/ViewController.m
@@ -8,8 +8,9 @@
 
 #import "ViewController.h"
 #import "LLDebug.h"
-#import <AFNetworking.h>
-#import "LLURLProtocol.h"
+
+// Used to example.
+#import "NetTool.h"
 
 static NSString *const kCellID = @"cellID";
 
@@ -24,8 +25,10 @@ static NSString *const kCellID = @"cellID";
 
 - (void)viewDidLoad {
     [super viewDidLoad];
+    
     // LLDebugTool need time to start.
     sleep(0.5);
+    __block __weak typeof(self) weakSelf = self;
     
     if ([[NSUserDefaults standardUserDefaults] boolForKey:@"openCrash"]) {
         [[NSUserDefaults standardUserDefaults] setBool:NO forKey:@"openCrash"];
@@ -43,17 +46,22 @@ static NSString *const kCellID = @"cellID";
         dispatch_async(dispatch_get_main_queue(), ^{
             if (!connectionError) {
                 UIImage *image = [[UIImage alloc] initWithData:data];
-                self.imgView.image = image;
+                weakSelf.imgView.image = image;
             }
         });
     }];
-
+    
+    
     // Json Response
-    [[AFHTTPSessionManager manager] GET:@"http://baike.baidu.com/api/openapi/BaikeLemmaCardApi?&format=json&appid=379020&bk_key=%E7%81%AB%E5%BD%B1%E5%BF%8D%E8%80%85&bk_length=600" parameters:nil progress:nil success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
+    [[NetTool sharedTool].afHTTPSessionManager GET:@"http://baike.baidu.com/api/openapi/BaikeLemmaCardApi?&format=json&appid=379020&bk_key=%E7%81%AB%E5%BD%B1%E5%BF%8D%E8%80%85&bk_length=600" parameters:nil progress:nil success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
+        
     } failure:^(NSURLSessionDataTask * _Nullable task, NSError * _Nonnull error) {
+        
     }];
     
-    // Log
+    // Log.
+    // NSLocalizedString is used for multiple languages.
+    // You can just use as LLog(@"What you want to pring").
     LLog(NSLocalizedString(@"initial.log", nil));
 }
 
@@ -62,7 +70,7 @@ static NSString *const kCellID = @"cellID";
     NSString *url = @"http://baike.baidu.com/api/openapi/BaikeLemmaCardApi?&format=json&appid=379020&bk_key=%E7%81%AB%E5%BD%B1%E5%BF%8D%E8%80%85&bk_length=600";
     
     // Use AFHttpSessionManager
-    [[AFHTTPSessionManager manager] GET:url parameters:nil progress:nil success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
+    [[NetTool sharedTool].afHTTPSessionManager GET:url parameters:nil progress:nil success:^(NSURLSessionDataTask * _Nonnull task, id  _Nullable responseObject) {
         [[LLDebugTool sharedTool] showDebugViewControllerWithIndex:0];
     } failure:^(NSURLSessionDataTask * _Nullable task, NSError * _Nonnull error) {
         [[LLDebugTool sharedTool] showDebugViewControllerWithIndex:0];
@@ -70,9 +78,8 @@ static NSString *const kCellID = @"cellID";
     
     // Use AFURLSessionManager
     /*
-    AFURLSessionManager *manager = [[AFURLSessionManager alloc] init];
     NSURLRequest *request = [[NSURLRequest alloc] initWithURL:[NSURL URLWithString:url]];
-    [manager dataTaskWithRequest:request uploadProgress:nil downloadProgress:nil completionHandler:^(NSURLResponse * _Nonnull response, id  _Nullable responseObject, NSError * _Nullable error) {
+    [[NetTool sharedTool].afURLSessionManager dataTaskWithRequest:request uploadProgress:nil downloadProgress:nil completionHandler:^(NSURLResponse * _Nonnull response, id  _Nullable responseObject, NSError * _Nullable error) {
         [[LLDebugTool sharedTool] showDebugViewControllerWithIndex:0];
     }];
      */
@@ -89,10 +96,9 @@ static NSString *const kCellID = @"cellID";
 
 - (void)testHTMLNetworkRequest {
     //NSURLSession
-    NSURLSession *session = [NSURLSession sharedSession];
     NSMutableURLRequest *urlRequest = [[NSMutableURLRequest alloc] initWithURL:[NSURL URLWithString:@"https://www.baidu.com"]];
     [urlRequest setHTTPMethod:@"GET"];
-    NSURLSessionDataTask *dataTask = [session dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
+    NSURLSessionDataTask *dataTask = [[NetTool sharedTool].session dataTaskWithRequest:urlRequest completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
         [[LLDebugTool sharedTool] showDebugViewControllerWithIndex:0];
     }];
     [dataTask resume];