+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.0.3](https://github.com/HDB-Li/LLDebugTool/releases/tag/1.0.3) (05/31/2018)
+
+Fix some leaks.
+
+#### Update
+
+* Call `CFRelease` in `LLAppHelper`.
+* Resolve circular references caused by the `NSURLSessionDelegate` in `LLURLProtocol`.
+* Call `Free` in `LLBaseModel`.
+* Fix analyze warning in `LLBaseViewController` / `LLFilterOtherView`.
+* Uncoupled code in `LLTool` / `LLNetworkContentVC`.
+
+#### Additional Changes
+
+* Add `NetTool`(Use URLSession in a singleton.) and update `ViewController` (Fix a circular reference.)
+
 ## [1.0.2](https://github.com/HDB-Li/LLDebugTool/releases/tag/1.0.2) (05/21/2018)
 
 * Fix the side gesture recognizer bug when pop.